requirements.txt
pandas
python-dotenv
openai
langchain
tree-of-thoughts